only upload existing files

+++ add_files_to_branch.rb
@@ -21,21 +21,25 @@ branch = branch_prefix.empty? ? id.to_s : "#{branch_prefix}.#{id}"
 jats_uploaded_path = "#{branch}/10.21105.#{branch}.jats"
 crossref_uploaded_path = "#{branch}/10.21105.#{branch}.crossref.xml"
 
-crossref_gh_response = github_client.create_contents(papers_repo,
-                                            crossref_uploaded_path,
-                                            "Creating 10.21105.#{branch}.crossref.xml",
-                                            File.open("#{crossref_path.strip}").read,
-                                            branch: branch)
+if File.exist?(crossref_path)
+  crossref_gh_response = github_client.create_contents(papers_repo,
+                                              crossref_uploaded_path,
+                                              "Creating 10.21105.#{branch}.crossref.xml",
+                                              File.open("#{crossref_path.strip}").read,
+                                              branch: branch)
 
-system("echo '::set-output name=crossref_html_url::#{crossref_gh_response.content.html_url}'")
-system("echo '::set-output name=crossref_download_url::#{crossref_gh_response.content.download_url}'")
+  system("echo '::set-output name=crossref_html_url::#{crossref_gh_response.content.html_url}'")
+  system("echo '::set-output name=crossref_download_url::#{crossref_gh_response.content.download_url}'")
+end
 
 
-jats_gh_response = github_client.create_contents(papers_repo,
-                                            jats_uploaded_path,
-                                            "Creating 10.21105.#{branch}.jats",
-                                            File.open("#{jats_path.strip}").read,
-                                            branch: branch)
+if File.exist?(jats_path)
+  jats_gh_response = github_client.create_contents(papers_repo,
+                                              jats_uploaded_path,
+                                              "Creating 10.21105.#{branch}.jats",
+                                              File.open("#{jats_path.strip}").read,
+                                              branch: branch)
 
-system("echo '::set-output name=jats_html_url::#{jats_gh_response.content.html_url}'")
-system("echo '::set-output name=jats_download_url::#{jats_gh_response.content.download_url}'")
+  system("echo '::set-output name=jats_html_url::#{jats_gh_response.content.html_url}'")
+  system("echo '::set-output name=jats_download_url::#{jats_gh_response.content.download_url}'")
+end
